user_auth/pyproject.toml
[project]
name = "user-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "simple-salesforce>=1.12.9",
]
